Replaced ResponseError instances with BadRequestException + put all logic on service + differentiated PUT and PATCH instance

+++ library/src/main/java/com/gisdev/library/service/BookService.java
@@ -1,8 +1,10 @@
 package com.gisdev.library.service;
 
+import com.gisdev.library.dto.ResponseError;
 import com.gisdev.library.dto.request.BookCreateDTO;
 import com.gisdev.library.dto.request.BookUpdateDTO;
 import com.gisdev.library.entity.Book;
+import com.gisdev.library.exception.BadRequestException;
 import com.gisdev.library.repository.BookRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -23,8 +25,11 @@ public class BookService {
         return bookRepository.existsById(id);
     }
 
-    public Book createBook(BookCreateDTO request) {
+    public Object createBook(BookCreateDTO request) {
 
+        if (titleExists(request.title())) {
+            return new BadRequestException("Book with this title already exists");
+        }
         Book book = Book.builder()
                 .title(request.title())
                 .author(request.author())
@@ -37,49 +42,34 @@ public class BookService {
         return bookRepository.save(book);
     }
 
-    public Book updateBook(Long id, BookUpdateDTO request) {
+    public Object updateBook(Long id, BookUpdateDTO request) {
 
         Book book = bookRepository.findById(id).orElse(null);
-
         if (book == null) {
-            return null;
+            return new BadRequestException("Book with this id does not exist");
         }
-
-        if (request.title() != null) {
-            book.setTitle(request.title());
-        }
-        if (request.author() != null) {
-            book.setAuthor(request.author());
-        }
-        if (request.genre() != null) {
-            book.setGenre(request.genre());
-        }
-        if (request.section() != null) {
-            book.setSection(request.section());
-        }
-        if (request.price() != null) {
-            book.setPrice(request.price());
-        }
-        if (request.yearOfPublication() != null) {
-            book.setYearOfPublication(request.yearOfPublication());
-        }
+        book.setTitle(request.title());
+        book.setAuthor(request.author());
+        book.setGenre(request.genre());
+        book.setSection(request.section());
+        book.setPrice(request.price());
+        book.setYearOfPublication(request.yearOfPublication());
 
         return bookRepository.save(book);
     }
 
-    public void deleteBook(Long id) {
+    public Object deleteBook(Long id) {
 
+        if (!idExists(id)) {
+            return new BadRequestException("Book with this id does not exist");
+        }
         bookRepository.deleteById(id);
+        return new ResponseError("Deletion successful");
     }
 
     public List<Book> getAllBooks() {
 
         return bookRepository.findAll();
-    }
-
-    public Book getBook(Long id) {
-
-        return bookRepository.findById(id).orElse(null);
     }
 
     public List<Book> getAllbooks() {
